dev: round Float32 to after point

+++ periphery/en6400/lis3dhtr_windows.go
@@ -22,9 +22,9 @@ import "fmt"
 
 // AccelerationData 存储加速度数据
 type AccelerationData struct {
-	X float64
-	Y float64
-	Z float64
+	X float32
+	Y float32
+	Z float32
 }
 
 // To String
